Made a ton of progress, but this is a mess.
I might go back and update this one day so it's actually functional. There's way too much repeating, I thought that would be better but it's really a nightmare. I think I could make this one function with a flag for line. Depending on the val, we could trigger different things. That might be the meta, but it's 3:30 in the morning and I feel like I just got done climbing a mountain.